created the middleware which authenticate and authorize user
