+++ src/routes/user.route.ts
@@ -7,6 +7,7 @@ import { isBlocked } from '../middleware/isBlockedMiddleware.js';
 import { userAuthJWT, sellerAuthJWT, adminAuthJWT, verifyToken } from '../middleware/verfication.middleware.js';
 import { isRoleAdmin } from '../middleware/checkAdminRoleMiddleware.js';
 import { verifyTwoFactor } from '../controllers/userController/2Factor.controller.js'
+import authenticateAndAuthorizeUser from '../middleware/authenticateAndAuthorizeUser.js'
 
 export const usersRouter = express.Router();
 
@@ -108,7 +109,7 @@ usersRouter.post('/verify', verifyToken, UserController.verifyUser);
  * @param {import('express').Response} res - The Express response object.
  * @returns {Promise<void>} A Promise that resolves when the login operation is complete.
  */
-usersRouter.post('/login/user', isBlocked, loginController.login);
+usersRouter.post('/login/user', isBlocked, authenticateAndAuthorizeUser, loginController.login);
 usersRouter.get('/users/:id', UserController.getUserById);
 usersRouter.get('/users',isRoleAdmin, UserController.getAllUsers);
 usersRouter.put('/users/update//:id', UserController.updateUser);
